Implement GraphQL call to fetch user data
