+++ src/components/episodeTracker.jsx
@@ -5,13 +5,6 @@ import { getProfile } from '../utils/auth';
 const EpisodeTracker = () => {
   const userStatus = useContext(context);
 
-  // Fetch from episode tracker API
-  useEffect(() => {
-    if (!userStatus.userEpisodes) {
-      userStatus.changeUserEpisodes([]);
-    }
-  });
-
   return (
     <p>Episode tracker</p>
   );
